ROBOT-1118: Continue to split up robot.py and rename
